Remove redundant SFB un-swizzle/re-swizzle and return BF16 directly
SFB_all is already per-expert swizzled because each expert is independently
quantized by quantize_nvfp4 (which swizzles each expert's scales separately).
The un-swizzle + re-swizzle round-trip was a no-op that wasted 12.5% extra
memory for the cached copy. Pass SFB_all directly to the fused C dispatch.

Also return BF16 from the grouped GEMM instead of converting to FP32.
CUTLASS already accumulates in FP32 and outputs BF16 — the .float()
conversion was an unnecessary copy.

+++ bitsandbytes/functional.py
@@ -1387,7 +1387,7 @@ def gemm_nvfp4_grouped(
         K: Input dimension per expert.
 
     Returns:
-        Output tensor of shape (total_tokens, N) in float32 with tensor scales applied.
+        Output tensor of shape (total_tokens, N) in bfloat16 with tensor scales applied.
     """
     num_experts = expert_offsets.numel() - 1
